+++ example_command_results/lib/homepage_toWidget.dart
@@ -34,7 +34,8 @@ class _HomePageState extends State<HomePage> {
           ),
           Expanded(
             // Handle events to show / hide spinner
-            child: ValueListenableBuilder<CommandResult<String?, List<WeatherEntry>>>(
+            child: ValueListenableBuilder<
+                    CommandResult<String?, List<WeatherEntry>>>(
                 valueListenable: weatherManager.updateWeatherCommand.results,
                 builder: (BuildContext context, result, _) {
                   return result.toWidget(
@@ -50,7 +51,8 @@ class _HomePageState extends State<HomePage> {
                       children: [
                         Text('An Error has occurred!'),
                         Text(result.error.toString()),
-                        if (result.error != null) Text('For search term: ${result.paramData}')
+                        if (result.error != null)
+                          Text('For search term: ${result.paramData}')
                       ],
                     ),
                   );
@@ -63,14 +65,18 @@ class _HomePageState extends State<HomePage> {
               children: <Widget>[
                 Expanded(
                   child: ValueListenableBuilder<bool>(
-                    valueListenable: weatherManager.updateWeatherCommand.canExecute,
+                    valueListenable:
+                        weatherManager.updateWeatherCommand.canExecute,
                     builder: (BuildContext context, bool canExecute, _) {
                       // Depending on the value of canEcecute we set or clear the Handler
-                      final handler = canExecute ? weatherManager.updateWeatherCommand : null;
+                      final handler = canExecute
+                          ? weatherManager.updateWeatherCommand
+                          : null;
                       return ElevatedButton(
                         child: Text("Update"),
                         style: ElevatedButton.styleFrom(
-                            primary: Color.fromARGB(255, 33, 150, 243), onPrimary: Color.fromARGB(255, 255, 255, 255)),
+                            primary: Color.fromARGB(255, 33, 150, 243),
+                            onPrimary: Color.fromARGB(255, 255, 255, 255)),
 
                         /// because of a current limitation of Dart
                         /// we have to use `?.execute` if the command is