+++ example_command_results/lib/listview.dart
@@ -14,10 +14,12 @@ class WeatherListView extends StatelessWidget {
         leading: data[index].iconURL != null
             ? Image.network(
                 data[index].iconURL!,
-                frameBuilder: (BuildContext context, Widget child, int? frame, bool wasSynchronouslyLoaded) {
+                frameBuilder: (BuildContext context, Widget child, int? frame,
+                    bool wasSynchronouslyLoaded) {
                   return child;
                 },
-                loadingBuilder: (BuildContext context, Widget child, ImageChunkEvent? loadingProgress) {
+                loadingBuilder: (BuildContext context, Widget child,
+                    ImageChunkEvent? loadingProgress) {
                   if (loadingProgress == null) return child;
                   return CircularProgressIndicator();
                 },