+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2.0.0] - 03.03.2021
+
+* finished null safety migration
+* thrownExceptions only notifies its listeners now when a error happens and not also when it is reset to null at the beginning of a command
+
 ## [1.0.0-nullsafety.1] - 15.02.2021
 
 * Added `toWidget()` extension method on `CommandResult`